Added missing semicolons. Hid the password and added a button with animation to show it. Fixed client routes. Fixed the Loader at the EditPage.
Added missing semicolons to avoid eslint warnings. Hid the password at the HomePage and added an action-btn to show it. It can be done in 2 ways by clicking or hovering and waiting the button. After the password is shown you can lock it. Button icons are animated when changed. This is done using a custom hook called animatedIcon.hook.js. Fixed client side routes to redirect properly when the user is logged in and not. Fixed the Loader component at the EditPage to be displayed. Fixed the credentials.loading property to be set to true before the request to the backend, so the Loader component will be dispayed at the right time and no errors cause by the data missing will not occure.